cloud_storage_e2e_fuzz_test: wait on eviction
In the fuzz test, we schedule GC ops, but currently wait some fixed
amount of time for completion. This makes the test unnecessarily slow,
since most GC will actually happen quite quickly.

This adds an interface to the log_eviction_stm to see when the last
processed eviction event was, and uses it in the fuzz test to wait until
completion (though a timeout is non-fatal, in case the GC actually
didn't schedule new work)..