archival: Enable segment merging for compacted ...
... partitions. Previously, the adjacent segment merging was enabled
only for non-compacted partitions. The compacted partitions are
reuploaded when the compaction kicks in. This was the reasoning behind
the decision. It turned out that with the compacted segment reupload we
may end up with large number of small segments if there are a lot of
leadership transfers or the segments are compacted a lot and reupload
engaged before the cross segment compaction was completed.

This commit modifies the adjacent segment merger. It can now modify
only compacted or only non-compacted segments. For compacted partitions
it won't be doing anything before the local segments are compacted.