storage: use B-Tree for compaction reducer
Swap out the node hash map for a B-Tree. The issue with the node hash
map is it's growth pattern. It doubles in size whenver it nears its
capacity, which results in large allocations. The B-Tree is a lot more
well behaved in this regard.

Note how the key into the B-Tree is the hash of the key. This avoids a
lot of expensive string compares and maintains the randomness for when
we evict from the index. Since there can be collisions, I've used the
multimap variant of the tree.

(cherry picked from commit b2bab8b4f74e2765fc2e9345a4db09efbb944830)