ducktape: create bucket before starting cluster
This is typically the correct order of deployment, though most tests
didn't really rely on this, since tests will wait for cluster creation
to complete before creating topics.

Now that the cluster metadata uploader exists and uploads early on after
starting up, we need remote storage to be up and kicking before the
cluster is completely up.