raft: suppress heartbeats with a guard
Previously, concurrency in heartbeat suppression requests was managed by
checking the sequence number of the request on the unsuppress call. This
protected against the following:
* suppress(seq=1)
* suppress(seq=2)
* unsuppress(seq=1)
* unsuppress(seq=2)

But didn't protect against
* suppress(seq=1)
* suppress(seq=2)
* unsuppress(seq=2)
(should still be suppressed here)
* unsuppress(seq=1)

The latter scenario is possible when we skip the recovering follower in
replicate_entries_stm. Fix this by introducing a RAII guard that takes
care of updating the in-flight requests count.