cluster: add producer_state and producer_state_manager
producer_state enscapsulates all the state of a producer producing
to a raft group. This state should _ideally_ include both idempotency
and transactional state but for the current patch series, we only
port idempotency related state which primarily includes sequence
number tracking. Porting of transactional state will be done as future
work.

Prior to this patch series, the idempotency state in the stm was tracked
across multiple hash maps keyed by a PID. These include

* inflight_requests session information
* seq_table map that tracks sequence numbers
* idempotency lock map

With the switch to producer_state, all this information is encapsulated
into a producer_state object for a given PID and in a more concise
manner.

producer_state_manager is a shard level instance that manages timely
eviction of excess/expired pids on a given shard. This decoples
eviction logic from stm. Prior to this patch there was an eviction
fiber and logic for each raft group that was responsbile for tracking
LRU-ness. That is now moved to single fiber that does it at shard leve.

Additionally eviction is implemented in a lock-free manner, more details
in the class header.