tests/cg_consistency: changed timeout to 120s
KgoVerifierConsumer uses join timeout of 60 seconds, this may lead to
situation in which consumer group will be rebalancing for 1 minute which
resulted in a test assertion to be triggered. Changed the timeout to
account for the consumer join timeout.

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit 08fd7aa09f2259948a20666b26db6d65cec78ae3)