ssx: add a work_queue utility
This is a small utility that manages running tasks sequentially on a
single fiber.

This is useful for simplifying control loops by executing everything on
a single fiber without need for locking/etc.

As a potenial future extension: for long running async work, this queue
could have the ability to manage spinning off background fibers to
handle work, then re-enqueuing the result of that work. The advantage of
having the work_queue track that is that the bookkeeping is
consolidated into a single place and we can cleanly handle shutdown.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>