kvstore::start: Defer snapshot save until after version check
This commit refactors the kvstore startup recovery sequence to
account for situations where an upgrade is halted due to an
incompatible logical version.

In order to assert logical version compatibility, we must first
materialize the key value store in memory. After doing so, we
can assert on the difference between the cluster's active version
and the latest logical version baked into the loaded binary.

However, at the end of the kvstore startup sequence, after replaying
any snapshots and/or segments into memory, we save a snapshot of the
result back to disk and GC the corresponding segments as needed. This
is done to avoid accumulating a bunch of tiny segments if there are
frequent restarts without much useful activity.

The result of all this is that, even when a new binary version has
been rejected on startup, that node's kvstore has been snapshotted
back to disk based on the _new_ version's serialization format.
This can lead to unexpected behaviors when the node must eventually
be "downgraded" to a compatible logical version.

So this commit adds code to defer the kvstore's opportunistic GC
operations until _after_ the version compatibility assertion has
been made, ensuring that the data directory is not modified when
a node upgrade is disrupted in this way.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>