kvstore::start: Defer kvstore GC until after version check
This commit refactors the kvstore startup recovery sequence to
account for situations where an upgrade is halted due to an
incompatible logical version, deferring opportunistic garbage
segment collection until _after_ the logical version check.

This was done to account for the fact that a failed upgrade of
this type requires a downgrade to a version compatible with the
rest of an active cluster but possibly _not_ forward compatible
with the newer version's kvstore.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>