transform: introduce transform manager
This is the main controller for processors. It ensures that there are
processors running on it's core for all of the leader ntp's that have
transforms attached as inputs.

In the case of failures we retry with backoff exponentially, this state
is tracked along side it's processor in a table that includes probes as
well (as these must be shared on a core if a core owns multiple
partitions for an input topic that has a transform).

The relationship of there being many transforms for many partitions
means that the logic here is non trivial of keeping track of all the
state, as well as tracking if a transform is running or not.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>