ducktape: ManyPartitionTest - Explicitly set background traffic bandwidth
So far we had not been giving a target bandwidth to the background
kgo-repeater traffic.

As a result the clients would just spam as fast as possible and make the
test flaky and unreliable as this might overload RP.

Especially on Arm this causes leadership to not stabilize.

Note we also change the final soak phase to run for two minutes instead
of a number of bytes and calculate this based on the expected bandwidth.
This again makes the test more predictable independent of where it's
running.

Fixes https://github.com/redpanda-data/redpanda/issues/10507