raft: add leader dirty_offset to protocol_metadata
Leader dirty_offset will be used by followers to distinguish between
ordinary and recovery append_entries and also to get a sense of how many
offsets are left to recover. For coordinated recovery, receiving an
append_entries request with prev_log_index == dirty_offset will signal
to the follower that recovery is over and another raft group can be
allowed to start recovery.