cloud_storage: debug log on query below log start
Previously, when the start offset of the log advanced while processsing
a fetch resulting in the fetch to land below the start of the log, we
would log at ERROR level. However, this situation is normal when the
consumer is fetching close to the boundary from which retention is
operating.

This patch detects such situations and downgrades the log to the DEBUG
level.

Fixes #12105

(cherry picked from commit cd1886fdbfd1c8ac5809a70adc4d50cb865d2625)