rptest: Add Z3 based model
The model contains set of expressions. Every expression corresponds to
some functionality in redpanda. The model describes relationships
between different expressions. E.g. ts_read is a segment upload, and
ts_delete is a segment deletion.

The commit also contains a generic interface for the test that uses the
mode. It implements things like topic property updates and log data
subscription.

The interface is consumed by a set of validators and inputs. Every input
modifies something in redpanda. For intance, the input may change
redpanda configuration parameter or alter a topic config. Validators are
checking some conditions by examining logs, metrics or content of the S3
bucket.

Evern expression creates a set of validators and/or inputs. The model
produces a set of test cases. Every test case contains a single solution
of the Z3 model and inputs and validators that set up the test and check
the conditions.

Currently, the model is not full. Some aspects are missing form it. For
instance, the time based retention is not defined. The test cases are
generated by a function 'get_tiered_storage_test_cases' which is
supposed to be used with 'matrix' decorator. It accepts one parameter
'fast_run'. If 'fast_run' is False the function will return test cases
for all possible solutions (84 as for now). If 'fast_run' is false it
will only add a series of curated test cases (9). This is a way to add
limited amount of work to the CI runs and also to disable flaky tests or
tests that hit bugs in Redpanda.