Report leaderless and URP count in HMB
In addition to an explicit list of leaderless partitions and under
replicated partitions, include also a count of each of these
lists, except that the lists are subject to truncation while
the count is the full count as if truncation did not occur.

Issue #11378.

(cherry picked from commit d64e5aab8231bd91e02fc37dff7100e63f0f00d0)