Bench for health report aggregation
In order to solve issue #11378, we will produce not-truncated counts
of leaderless and under-represented partitions. To do this, we need
to keep adding partitions to a set/map where we would not
previously have done so, in order to get a correct de-duplicated count
of these elements.

For very large numbers of partitions, this could be a performance
problem, so I wrote this benchmark.

Results for 100k partitions and 10 nodes (the config in the
benchmark):

test                         iterations      median          inst
original                            134     1.380ms     6400321.7
original_unlimited                   34    25.006ms   141876040.5
current                              66     7.363ms    42025980.6

Original is the code before the change, which stops adding elements
to the set when their size reaches 128. original_unlimited is the same
code but with the limit removed: the easiest way to get the counts,
while current is the actual code checked in the immediately prior
changes which uses the `collector` to collect per-topic lists of
partition IDs.

So the new code is about 5x slower than the old approach which didn't
count all partitions, but at least 3x fast than the direct approach to
counting the partitions. Importantly even for this very large test case
we are well below the reactor stall threshold (though well above the
default task quota). Overall the time of less than 10 ms seems fine
for a command which seems to take ~35ms even on a totally unloaded
single-node system over localhost.

(cherry picked from commit ced4a02474fee04e6a08ba3dfd3f98e2116f84ec)