Additional tests for health report aggregation
Unit tests for health report aggregation that we've changed
during this series.

Checks that the expected results are obtained in some simple scenarios
and also that list truncation occurs at the expected amount and that
the count is correct even in this case.

Issue #11378.

(cherry picked from commit 7fbd5328676a6a09461be02d04824f43d03665c1)