Add the count of certain partitions to admin API
This change adds the count of leaderless and under-replicated
partitions to the cluster_health_overview API. This API already
contained an explicit list of such partitions, so what help is the
count? Well both lists are truncated at 128 entries, but the count
reports the true value.

This can be shown in the rpk health command so that progress can be
monitored even when the lists are beyond the truncation limit.

Fixes #11378.