archival: Fix adjacnet segment merger state transition
If the configuration parameter that enables ASM is OFF and the partition
becomes a leader we will disable ASM. Then, even if the config is turned
ON the ASM will stay disabled until the next leadership transfer. This
is because we're checking the config to decide should we enable ASM job
or not. But the ASM job checks config internally so config state
shouldn't contribute to its state.

This commit fixes the behavior.