archival: add boilerplate scrubber job
This commit introduces a stub housekeeping scrubber job. At this point
it includes the scaffolding required by the housekeeping service and the
scheduling logic. A future commit will plug in the actual scrubbing.

A scrubbing scheduling utiliy class is also added. Separating this logic
from the scrubber itself allows for writing unit tests.