cloud_storage: store anomalies in partition manifest
This commit extends the partition manifest to include the anomalies
detected by the scrubber. The next commit will add the "write path" for
this. Note that the anomalies are not included in the serde format, so
they will not be uploaded to the cloud. They are, however, included in
the snapshot.

The anomaly validation logic is also included in this commit (see
partition_manifest::process_anomalies). This is where false positives
detected by the scrubber are removed.