rename delete_retention_ms to log_retention_ms.
this commit is mostly a rename, since semantically the property is
already used as log.retention.ms inside redpanda.
log_retention_ms is the time based retention of the whole (partition)
log.
delete_retention_ms was meant to mimic delete.retention.ms in
kafka, and that property deals with retention for a specific type of
records.