storage/segment_index: force use of ss::lowres_system_clock
seastar::lowres_system_clock is meant to be a clock compatible with
std::system_clock, while being cheaper to call now().

this commit changes segment_index::maybe_track to nudge the caller to
use this clock. in the backend model::timestamp is still used
as storage for compatibility with the pre-existing functions
and stability of the byte representation.