rptest: increase purge grace period in racy test
`remote_partition::finlalize` may do a final manifest upload if
the uploaded manifest is not at the latest insync offset. This may race
with the removal of that manifest by the purger. Note that this can only
happen if the grace timeout is set super low (this test had it at 5ms).
Not an issue in the field, but let's bump the grace period to avoid
races in the test.

(cherry picked from commit 204e5229973ba7491f99f40c2b5324c1cb8fd3eb)