c/partition_balancer: take reclaimable partition size into account
With the introduction of space management data eviction becomes lazy:
partitions can grow beyond their eviction limit if there is sufficient
space to do it and only if disk space pressure becomes high, this
"extra" data is evicted.

For partition balancing this means two things: 1) partitions sizes can
vary widely between replicas and 2) when predicting how much space the
partition will take on a node if it is moved there, we must take into
account the possibility of reclaim (otherwise we won't allow moves that
are perfectly possible if reclaim is applied).

So this commit we 1) store current sizes in a per-replica map instead of
a single value and 2) use non-reclaimable size when predicting the
partition size on the target replica.