cloud_storage: Track mem in the materialized_resourcs
Instead of tracking individual segments/readers/partition-readers using
three different semaphores track all these objects using one semaphore
called _mem_units. The semaphore's counter contains number of bytes
returned by memory_groups::tiered_storage_max_memory().

Instead of acquiring single unit the code computes projected object size
and then uses this value to acquire a semaphore units. The size of the
segment reader is calculated based on readahead and buffer size. The
size of the remote partition reader is estimated using constant and the
size of the remote_segment object is also estimated using another
constant. When the actual remote_segment is created and registered the
code adjucts the units by releasing some of the if estimate overshoot or
by taking units if the estimate was too low. In the later case the
remote_segment object marked using 'overcommit' flag. The eviction
prioritizes objects with this flag set to true.

Parameters that define max number of segments/readers in the cloud
storage are ignored in this commit.