cloud_storage: Limit parallel segment downloads
Use a dedicated semaphore to avoid draining the network. Previously it
was limited by 'cloud_storage_max_partition_readers_per_shard'
configuration parameter. This is not a good solution because there is
still a way for a tiered-storage to download more than
'cloud_storage_max_partition_readers_per_shard' segments in parallel.
This commit introduces a dedicated semaphore to the
materialized_resources. This semaphore is initialized using the
'cloud_storage_max_partition_readers_per_shard' value and is acquired in
the 'remote' during the segment download. The default value for the
'cloud_storage_max_partition_readers_per_shard' is
'cloud_storage_max_connections / 2'. If the parameter exceeds
'max_connections' it will be clamped.