cloud_storage: Update memory limit dynamically
Use 'cloud_storage_max_segment_readers_per_shard' to set the memory
limit for tiered-storage. If this property is not null the
tiered-storage will caclulate memory requirements for that amount of
readers and use this value as a memory limit. The value can't go above
the limit in memory groups.

Another change is that readahead and buffer size parameters can now
affect the memory limit because reader memory untilization depends on
them. So when this parameters are changing the limit will be
recalculated.

The parameter 'cloud_storage_max_partition_readers_per_shard' is used to
set max read-path download parallelism. Its meaning is not number of
partition readers but number of parallel downloads. The name stayed the
same to limit the impact on existing clusters. But we should introduce
new property and add an alias in the future.

The 'cloud_storage_max_materialized_segments_per_shard' is depricated
because we limit things in the cloud storage by memory now.