Assume that ext filesystems are ext4
Due to scylladb/seastar#251 all of ext2, ext3 and ext4 will be
detected as ext2 (they shared the same superblock magic number).

This leads to an erroneous warning claiming that you are not running
on ext4 nor XFS. For now, we will simply that assume any ext detection
means ext4, as ext2 and 3 use is very limited as ext4 was rolled out
more than 15 years ago and almost entirely supplanted use of the
earlier variants.

In any case, the change only changes the warning message emitted, not
any other behavior.

Issue #13469.