Add ext4 to the syschecks FS check
We semi-support ext4 but in the syscheck we will now check for both
ext2 and ext4, as currently ext4 is always detected as ext2, but
in the future if ext4 is returned we should also treat this as ext4.

Issue #13469.

(cherry picked from commit 051c63c77b4a5823468c84554e513bfca9506da2)