cluster/topics_frontend: Move topic preprocessing into create_topics
Adds `shadow_indexing` and `cleanup_policy_bitflags` default fallback to
`create_topics`.

Also plumbs the `metadata_cache` down into `topics_frontend` for this
purpose, and in case we need to add more property fallbacks in the future.

This commit also relaxes an assertion in a fixture test [1] to account for
some additional requests during topic recovery when the test topic has
cloud storage switched on.

[1] `cloud_storage_rpfixture::topic_recovery_service_test.recovery_with_existing_topic`