cloud_storage: remember whether segment didn't have a tx index
`may_have_tx_index` can be used to avoid sending unnecessary requests to
the object storage. The field is used only as an optimization during the
GC process.

Note that although round-tripping between `segment_meta` and
`lw_segment_meta` losses the optimization, correctness is not violated.
This is acceptable as otherwise we would need to introduce a new field
in the metadata cstore. Or, store the `metadata_size_hint` in the
lightweight meta structure.