archival: refactor archive gc to allow reusing it
Refactor the batching out of the `garbage_collect_archive` into
`delete_segments` method so that it can be reused in other places.

Pass `lw_segment_meta` instead of raw object keys to allow logging the
segment name, counting the number of segments deleted. On one hand we're
copying extra structures, on the other we do late string
materialization.