cloud_storage: add logic for segment metadata scrub
This commit introduces a new utility function, `scrub_segment_meta`,
which compares a `current` segment meta with the previous one in order
to detect anomalies such as offset gaps, overlaps or bad delta offsets.

The placement of this function may seem odd, but it allows a future
commit to use it in order to validate new segments added to the
manifest.