cloud_storage: plug in segment meta scrubbing
This commit uses the `scrub_segment_meta` function added previously in
order to compare each pair of adjacent segments and detect anomalies.

We scrub from the newest manifest to the oldest. Within a manifest, the
scrub starts from the oldest segment since the column store is encoded
and can't support reverse iteration. Due to this ordering, some extra
logic is required in order to scrub segment metadata across manifest
gaps.