compaction: do not discard control batches
Currently we always discard control batches. This is a problem when the
batch being discarded is the last batch in a segment. In some edgecases
where there is no user data after the control batch (empty new segment),
consumers may be stuck behind LSO and repeatedly trying to poll for data
when there is none. This may look like an infinite loop from a consumer
POV but its just that there is no data after that offset as it got
compacted away. This situation is not possible in a regular self
compaction becuase we always retain the last batch.

With this commit we use regular compaction process for control batches,
so the resulting segment may end up with some of them and guaranteed to
retain the last batch. These are just filtered away on the client side
but gurantees the consumed offset makes progress.