r/consensus: track not flushed bytes
Added tracking the amount of bytes that are not flushed instead of
simply indicating that a part of the log is not flushed. This will allow
Redpanda to make a decision based on the amount of not flushed data
size.

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit de49f0c36617f33de76e92a09497e7c653d7b4dc)