storage/log_manager: apply_segment_ms for to while transformation
this is done to sidestep iterator invalidation that could happen during
the suspension point while co_awaiting for apply_segment_ms.

to do so, the while loops until it reaches a log that is already marked
with bflags::lifetime_checked. in the loop, current log is marked, list
is rotated for the next loop, and apply_segment_ms is awaited.