c/pb_backend: schedule tick immediately if health report is available
It may happen that the health report notification for a given node is
processed before the notification about the node addition. In this case
`partition_balancer_backend` would wait for another tick to trigger
rebalancing. Added immediate tick trigger when node add notification is
processed and health report for a node is already present.

Fixes: #12220

Signed-off-by: Michal Maslanka <michal@redpanda.com>