ducktape: port _wait_until_no_leader() to use partition debug state
In it's current form the method relies on health report to find
leaderless partitions. That doesn not work for rf=1 when the broker is
down because health report relies on an active replica reporting a
missing leader which is not possible in the above situation. Rewrite
using partition debug state can handle this scenario.