cloud_storage: add batching support inside remote::delete_objects
`delete_objects` splits the keys into batches and issue concurrently
properly sized requests depending on the backend.

After this change `delete_objects` method is expected to be called with
larger number of values. To be friendly to the Seastar memory allocator
we want to change this method to accept a `std::deque` rather than a
contiguous vector.

Adding a generic method with explicit template instantiation to support
gradual migration of the call-sites.