k8s: Change the integration test logic
In the case when broker is removed from the mock admin API the controller is failing always as Pod can not set up annotation with Node ID. The decommission is never called, so that StatefulSet is not updated. Integration tests are stuck due to that.

(cherry picked from commit 556b97be00f2486843ff0b28e630e5d69d9a1e20)