rpk: support producing using Schema Registry
This commit introduces the support for encoding
records in rpk topic produce using schema
registry.

rpk will query the schema from the registry and
encode according to the retrieved schema type.
The message then will be sent encoded to the
broker.

A following PR will handle the decoding
aspects in `rpk topic consume`