net: shutdown connections only after processing in-flight accepts
server is shutdown in 2 phases. A synchronous call to `shutdown_input`
where the listening sockets are closed followed a call to
`wait_for_shutdown` and waiting on the return future. This leaves an
opportunity for a race condition to occur.

Specifically, when the following sequence of events is observed:

1. `server::accept_finish` coroutine is called and suspended just before
   a new connection is added to the `server::_connections` list.

2. `server::shutdown_input` is called, connections from
   `server::_connections` are closed. Note that the incoming connection
   caused by the event above isn't yet added to that list.

3. `server::accept_finish` coroutine is resumed, the
   `server::_conn_gate` is open and the connection is serviced as usual.

4. `server::wait_for_shutdown` is called, and is waiting on
   `_conn_gate.close`. Eventually our connection is aborted, but not
    timely.

To break the race introduce an accept gate which is held while the
per-connection accept sequence is executed and until the connection is
passed over to the connection gate. `server::shutdown_input` now closes
the listener and waits for the accept gate to close ensuring that there
are no in-flight accepts. Only after that per-connection inputs are
closed too.

The `_as.check()` call in `server::accept_finish` method is just an
optimization after this change and not required for correctness.