partition_balancer: make maintenance mode opaque to the balancer
Maintenance mode (trigers re-election to move leadership from the
partitions hosted on the node in maintenance mode) doesn't do any-
thing which couldn't happen naturally in the cluster when the mode
is off.

So the balancer can't be affected if it ignores it.

(cherry picked from commit 124b89fa1f8004f973c4dc8db356a3940cdae760)