cluster: always create the log eviction stm
When topics were created with `cleanup.policy=compact` we wouldn't
allocate the `log_eviction_stm` as it wouldn't have any work to do
anyway. However, if the `cleanup.policy` is changed to `delete` or
`compact,delete` we still wouldn't allocate it. This leads to a bug
where trim-prefix command isn't working and neither topic retention
settings are applied until a node is restarted.

Since we don't have a way to enable/disable STMs at runtime, this commit
enables the log eviction stm unconditionally and checks the actual topic
properties whether trim-prefix is support or not.

Fixes https://github.com/redpanda-data/redpanda/issues/13588