cluster_recovery: fail recovery on error in same term
This updates the recovery backend to fail the recovery if an operation
fails (download fails, controller frontend command fails) during
recovery. This avoids a situation where the backend is caught
continuously trying to retry a recovery that is deterministically going
to fail (e.g. because some files are missing in S3).