s/snapshot: introduced `file_snapshot_writer`
Introduced a dedicated type being a simple `storage::snapshot_writer`
enriched with file name. The `snapshot_writer` was designed as a generic
class accepting only `ss::output_stream` without assuming that the
output stream is related with a certain file. Now the `snapshot_writer`
can be use in general context while `file_snapshot_writer` is intended
to be used where a snapshot is stored in a file.

Signed-off-by: Michal Maslanka <michal@redpanda.com>