cloud_storage: Match requests using URL
In topic recovery tests some assertions simply wait for a count of
requests made to S3 imposter. This may cause tests to fail when requests
not related to the test are observed by the imposter, throwing off the
expected count.

This change adds an optional predicate to the assertion so that the
request count can be matched after filtering for specific requests by
URL, method etc.