r/configuration: make configuration serde serializable
Added support for serde serialization in raft::configuration.
To determine if configuration should use `serde` or `adl` serialization
we peek for the configuration version. In serde version it will be
version from envelope while in adl serialized object it is going to be
an object version. This way we will be able to deserialize older
configuration types.

Signed-off-by: Michal Maslanka <michal@redpanda.com>