cluster/config_manager: fix double apply of configs
previously, to support validation, the code would call a method and then
call it again with a smp::invoke_on_all.

this pr preserve this behavior but uses invoke_on_others to eliminate
the double-apply on the current shard.

to support this, the lambdas needs to be no_except_copy/move, so the
parameters are taken by pointer, since they live in the parent scope