rptest: extend log allow list in scrubbing test
The test removes a segment from the manifest for the scrubber to detect.
A reupload may be attempted somewhere after the base offset of the
removed segment. The STM manifest will refuse to apply the command and
log an error. This commit updates the log allow list to include said
error.

Fixes #13732