archival: Audit meta in adjacent merger uploads
The 'ntp_archiver_service::upload' method is used by the adjacent
segment merger. It's an api for external services that need to upload
segments. The problem is that this method doesn't utilize the segment
stats to validate segment metadata. Normal upload path uses these stats
to check if the replicated metadata is correct.

This commit adds similar check to the 'upload' method.