ductape: bump timeout in force_reconfiguration test
Post restart, the test was timing out on achieving stable leadership
because of tight timeout. Restarted nodes are not seeing fresh health
reports with updated leadership information. Since health reports are
refreshed every 10s, we need to bump the wait timeout to factor in
multiple health reports incase the first one is stale.