rm_stm: interface for highest producer_id
Across recoveries of a cluster, the next producer_id used by the
id_allocator needs to be reset to avoid using any that may have been
used previously. So, this commit begins tracking the highest seen so
far, and exposes it for use by the upload path (to come in a subsequent
commit).