cloud_storage: Log trim failure based on file type
When an exhaustive trim fails, a common reason for this is .part files
which were identified to be deleted but during the trim they were
renamed to chunk/segment files.

This change adds a flag to the trim result, if the trim fails to free
enough space and a tmp file was missed during the trim, the log is
downgraded to INFO.

Additionally we catch filesystem error separately and if the error
caught during a trim is not an FS error, it is logged as an error.

An FS error not related to a temp file is also logged as an error.