serde: fix envelope operator overload foot-gun
Previously, if a sub-class of envelope did not define its own equality
operator, the comparison would ignore all state from the derived class
and use the defaulted comparison operator of `serde::envelope`. This
caused comparison to behave in unexpected ways.

The goal of this patch is to: force serde types to overload their
comparison operators iff they wish to be comparable. To this end, the
implementation of said operators in `serde::envelope` is templated in
order to distinguish implict calls when the operator is generated for
the base class and accidental calls from sub classes that didn't
implement their operators. The spaceship operator covers all possible
comparison operators apart from equality.

In order to get this change to compile, I also had to default the
comparison operator for a number of existing serde types.

(cherry picked from commit 7d37fa3c66764078e02cc0bdb0aefa13c34059c6)