admin_server: map `missing_node_rpc_client` to service unavailable
It may sometimes happen that the admin API request will be processed while
Redpanda is shutting down or it is in the middle of reconfiguration with
stale metadata. In this case an RPC client to currently reported leader
may no longer be available. In this case we should request client to
retry.

Fixes: #13855

Signed-off-by: Michal Maslanka <michal@redpanda.com>