compat: handle non-comparable types in compat tests
Previously, compat tests had an implicit requirement that the types
being tested are comparable. This doesn't always make sense. If the type
is not comparable, we round-trip it to serde and check the equality of
the "wire format".

(cherry picked from commit e392c44ab53ba8fc78e9a0ea428602f457f9aa72)