archival: handle same term syncs in archival STM
Previously, the archival STM used `persisted_stm::sync` in order to
ensure that its log is up to date before issuing new commands. This
works fine for leadership changes since they imply a term change.

However, the archiver may need to sync during the same term. We've seen
scenarios like this:

> Upload loop picks 4 candidates, uploads to cloud storage and
replicates a batch with 4 new segments. After the replication command
succeeds, the topic retention is updated. This re-starts the archiver.
When the archiver restarts it picks the same 4 segments and replicates
the addition commands again. Finally, the partition manifest detects
this, refuses the commands and prints the error log.

The previous sync method does not support such cases sync they happen in
the same term. To fix this, the archival STM now implements its own sync
method. If it detects that the term has changed, it will maintain the
pre-existing behaviour: sync up to the latest term. Otherwise, it will
wait for any ongoing replications to finish and then wait for the
commited offest to be applied.

Fixes #13753