wasm/cache: Add GC mechanism
In order to cleanup old factories we use a GC process that runs every 10
minutes.

It's a little overengineered - I don't expect to ever have more
factories than we can GC in a single task, but this will work even with
a large number of factories.

The advantage of this GC over some advanced caching is that we can keep
the compiled code around in memory if a transform is hopping nodes. It's
also simpler than trying to bookkeep an explicit ref count or something.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>