config: increase the cloud storage idle threshold
`cloud_storage_idle_threshold_rps` controls how many requests per second
are required in order for cloud storage to be considered idle. The
previous value was `1`, but that's too strict and leads to housekeeping
jobs not completing naturally. They'd eventually be drained, but it's
preferable to process them when we know the system is idle-ish than
waiting for the drain.

Also fixed a small issue where this config is compared to the timed
average from the sliding window. The average is over
`cloud_storage_idle_timeout_ms`, so we should multiply the raw rps
(requests per second value) by the number of seconds in the timeout.