test: fix Kafka usage tests
This test makes use of the comparator for `kafka::usage` which was
previously not explicitly defined. This caused the comparison to skip
all the members in the class. When the issues was fixed, the test
started failing and this commit fixes it.

Author: Rob Blafford
(cherry picked from commit be737e79331ec266d017e2b5332cb829c5c0a8f7)