cloud_storage: notify slow downs from the remote
This commit extends the notification type used by the remote to include
whether the request being notified was a retry or not. The
`upload_housekeeping_service` keeps track of a rolling average for the
total count of retries which happened on its shard. A later commit will
use this information to dinamically adjust the quotas used by the
background services managed in `upload_housekeeping_service`.