wasm: add configuration for async execution
The working model here is that wasmtime allocates a seperate stack to
execute Wasm on. That stack can be yielded to and will allow us to run
this on the reactor directly. Our host functions are also executed this
new stack, so now our host functions will only have 64 KiB to execute
on. We should add some test/debug support for ensuring that our host
functions do not use more stack memory than that. If more stack memory
is required, the option of making the host function async is possible,
as then the continuations will be executed on "normal" reactor thread
stacks, otherwise we'll have to consider a seperate memory pool for
stacks, as 128KiB is about the most we want to allocate at once in
Redpanda.

As another note these stacks are current mmap'd but there will be work
done in Wasmtime to plugin a seperate allocator in the runtime. That
will be a followup change to plug that in once it's ready.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>