wasm: Support registering async host functions
Async host functions are implemented via polling similar to Rust's
Future type. We implement this via a "done" boolean that is threaded
around and we mark it as true when the future completes. Wasmtime
ensures that all the arguments to the function will stay alive until our
polling function returns true.

There will be a followup where we will register the future with the
engine to not poll the status until future has completed.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>